+++ src/Catalog/Declination.php
@@ -53,6 +53,9 @@ final class Declination
     /** @var DeclinationOption[] */
     private $options;
 
+    /** @var CompanySummary */
+    private $company;
+
     /**
      * @internal
      */
@@ -77,6 +80,7 @@ final class Declination
         $this->options = array_map(static function (array $optionData) : DeclinationOption {
             return new DeclinationOption($optionData);
         }, $data['options']);
+        $this->company = new CompanySummary($data['company']);
     }
 
     public function getId(): string
@@ -187,5 +191,10 @@ final class Declination
          * by checking if the count of $variantIds is the same than the $declinationVariantIds'
          */
         return (count($variantIds) === $foundIds && count($variantIds) === count($declinationVariantIds));
+    }
+
+    public function getCompany(): CompanySummary
+    {
+        return $this->company;
     }
 }